refactor: add PushContactRepository
Getting data from DB was being treated on PushContactService.
Now, it should be obtained from PushContactRepository.
At moment, only moved the GetAllSubscriptionInfoByDomainAsync method,
being used by WebPushPublisherService.
