+++ Doppler.PushContact/Services/IPushContactService.cs
@@ -27,8 +27,6 @@ namespace Doppler.PushContact.Services
 
         Task<IEnumerable<string>> GetAllDeviceTokensByDomainAsync(string domain);
 
-        Task<IEnumerable<SubscriptionInfoDTO>> GetAllSubscriptionInfoByDomainAsync(string domain);
-
         Task<IEnumerable<string>> GetAllDeviceTokensByVisitorGuidAsync(string visitorGuid);
 
         Task<ApiPage<DomainInfo>> GetDomains(int page, int per_page);
